examples/document-processing/fraud-detection/resources/requirements.txt
PyPDF2==3.0.1
Pillow==11.0.0
